Tests updated for `{% embed %}` tag processing.
an `{% embed %}` block has been added to the input.html.twig with an additional gettext call, this is now picked up when the file is scanned, and tests have been updated accordingly.

The embed.html.twig file represents the embed template, it contains a gettext call, but it is not scanned automatically, instead it should be fed into gettext with all the other template files.

This design choice was made in order to keep the scanner simple as it only operates on the tokens rather than a full graph of dependencies. In fact I think the parse function doesn't even check for it's existence when parsing the tokens.